created the review model

+++ db/seeds.rb
@@ -1,13 +1,16 @@
-require 'json'
 require 'open-uri'
 
-Ingredient.destroy_all
+puts "Destroy ingredients"
+Ingredient.destroy_all if Rails.env.development?
 
-result = JSON.parse(open('https://www.thecocktaildb.com/api/json/v1/1/list.php?i=list').read)
-drinks = result['drinks']
+puts "Destroy Cocktails"
+Cocktail.destroy_all if Rails.env.development?
 
-drinks.each do |hash|
-  hash.each do |k, v|
-    Ingredient.create(name: v)
-  end
+puts "Create ingredients"
+url = "https://www.thecocktaildb.com/api/json/v1/1/list.php?i=list"
+ingredients = JSON.parse(open(url).read)
+ingredients["drinks"].each do |ingredient|
+  i = Ingredient.create(name: ingredient["strIngredient1"])
+  puts "create #{i.name}"
 end
+
